Refactor project structure by renaming main entry point and removing unused components
- Updated the main entry point from `src/main.tsx` to `src/app/main.tsx`.
- Deleted the `App.tsx`, `CourseCreationForm.tsx`, `CourseCard.tsx`, and other related components to streamline the codebase.
- Removed unnecessary imports and components to enhance maintainability and clarity.
