+++ src/app/App.tsx
@@ -1,0 +1,18 @@
+import { QueryClientProvider } from "@tanstack/react-query";
+import { queryClient } from "../lib/queryClient";
+import { Toaster } from "../components/ui/toaster";
+import { AuthProvider } from "../features/auth/context/AuthContext";
+import Router from "./Router";
+
+function App() {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <AuthProvider>
+        <Router />
+        <Toaster />
+      </AuthProvider>
+    </QueryClientProvider>
+  );
+}
+
+export default App;
