+++ src/features/ai-assistant/components/AIAssistant.tsx
