+++ src/features/ai-assistant/components/ChatInput.tsx
