+++ src/features/ai-assistant/components/ChatMessage.tsx
