+++ src/features/ai-assistant/components/ChatPanel.tsx
