+++ src/features/ai-assistant/components/ChatSidebar.tsx
