+++ src/features/course/components/CourseCard.tsx
