+++ src/features/course/components/CourseGrid.tsx
