+++ src/features/course/components/VideoPlayer.tsx
